--- src/main/java/qna/domain/Answers.java
@@ -1,27 +1,0 @@
-package qna.domain;
-
-import qna.CannotDeleteException;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-
-public class Answers {
-    private List<Answer> answers;
-
-    public Answers(List<Answer> answers) {
-        this.answers = Collections.unmodifiableList(answers);
-    }
-
-    public List<Answer> getAnswers() {
-        return answers;
-    }
-
-    public List<DeleteHistory> delete(User loginUser) throws CannotDeleteException {
-        List<DeleteHistory> deleteHistories = new ArrayList<>();
-        for (Answer answer : answers) {
-            deleteHistories.add(answer.delete(loginUser));
-        }
-        return deleteHistories;
-    }
-}
