+++ src/main/java/qna/domain/qna/ContentType.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.domain.qna;
 
 public enum ContentType {
     QUESTION, ANSWER;
