+++ src/main/java/qna/domain/qna/answer/Answer.java
@@ -1,8 +1,13 @@
-package qna.domain;
+package qna.domain.qna.answer;
 
 import qna.CannotDeleteException;
 import qna.NotFoundException;
 import qna.UnAuthorizedException;
+import qna.domain.deletehistory.DeleteHistory;
+import qna.domain.qna.AbstractEntity;
+import qna.domain.qna.ContentType;
+import qna.domain.qna.question.Question;
+import qna.domain.user.User;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
@@ -45,25 +50,12 @@ public class Answer extends AbstractEntity {
         this.contents = contents;
     }
 
-    public Answer setDeleted(boolean deleted) {
-        this.deleted = deleted;
-        return this;
-    }
-
     public boolean isDeleted() {
         return deleted;
     }
 
-    public boolean isOwner(User writer) {
-        return this.writer.equals(writer);
-    }
-
     public User getWriter() {
         return writer;
-    }
-
-    public String getContents() {
-        return contents;
     }
 
     public void toQuestion(Question question) {
