+++ src/main/java/qna/domain/qna/answer/AnswerRepository.java
@@ -1,6 +1,7 @@
-package qna.domain;
+package qna.domain.qna.answer;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import qna.domain.qna.question.Question;
 
 import java.util.List;
 import java.util.Optional;
