+++ src/main/java/qna/domain/qna/question/Question.java
@@ -1,11 +1,15 @@
-package qna.domain;
+package qna.domain.qna.question;
 
-import org.hibernate.annotations.Where;
 import qna.CannotDeleteException;
+import qna.domain.deletehistory.DeleteHistory;
+import qna.domain.qna.AbstractEntity;
+import qna.domain.qna.answer.Answer;
+import qna.domain.qna.answer.Answers;
+import qna.domain.qna.ContentType;
+import qna.domain.user.User;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 
 @Entity
@@ -20,10 +24,8 @@ public class Question extends AbstractEntity {
     @JoinColumn(foreignKey = @ForeignKey(name = "fk_question_writer"))
     private User writer;
 
-    @OneToMany(mappedBy = "question", cascade = CascadeType.ALL)
-    @Where(clause = "deleted = false")
-    @OrderBy("id ASC")
-    private List<Answer> answers = new ArrayList<>();
+    @Embedded
+    private Answers answers = new Answers();
 
     private boolean deleted = false;
 
@@ -41,24 +43,6 @@ public class Question extends AbstractEntity {
         this.contents = contents;
     }
 
-    public String getTitle() {
-        return title;
-    }
-
-    public Question setTitle(String title) {
-        this.title = title;
-        return this;
-    }
-
-    public String getContents() {
-        return contents;
-    }
-
-    public Question setContents(String contents) {
-        this.contents = contents;
-        return this;
-    }
-
     public User getWriter() {
         return writer;
     }
@@ -70,24 +54,11 @@ public class Question extends AbstractEntity {
 
     public void addAnswer(Answer answer) {
         answer.toQuestion(this);
-        answers.add(answer);
-    }
-
-    public boolean isOwner(User loginUser) {
-        return writer.equals(loginUser);
-    }
-
-    public Question setDeleted(boolean deleted) {
-        this.deleted = deleted;
-        return this;
+        answers = answers.add(answer);
     }
 
     public boolean isDeleted() {
         return deleted;
-    }
-
-    public List<Answer> getAnswers() {
-        return answers;
     }
 
     @Override
@@ -102,22 +73,22 @@ public class Question extends AbstractEntity {
         return deleteHistories;
     }
 
-    private DeleteHistory deleteQuestion(User loginUser){
+    private DeleteHistory deleteQuestion(User loginUser) {
         this.deleted = true;
         return DeleteHistory.of(ContentType.QUESTION, getId(), writer, LocalDateTime.now());
     }
 
     private List<DeleteHistory> deleteAnswers(User loginUser) throws CannotDeleteException {
-        return new Answers(answers).delete(loginUser);
+        return answers.delete(loginUser);
     }
 
     private void validateCanDelete(User loginUser) throws CannotDeleteException {
-        if(!isWriter(loginUser)){
+        if (!isWriter(loginUser)) {
             throw new CannotDeleteException("질문을 삭제할 권한이 없습니다.");
         }
     }
 
-    private boolean isWriter(User loginUser){
+    private boolean isWriter(User loginUser) {
         return this.writer.equalsNameAndEmail(loginUser);
     }
 }
