+++ src/main/java/qna/service/QnAService.java
@@ -6,11 +6,13 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import qna.CannotDeleteException;
 import qna.NotFoundException;
-import qna.domain.*;
+import qna.domain.deletehistory.DeleteHistory;
+import qna.domain.qna.answer.AnswerRepository;
+import qna.domain.qna.question.Question;
+import qna.domain.qna.question.QuestionRepository;
+import qna.domain.user.User;
 
 import javax.annotation.Resource;
-import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 
 @Service("qnaService")
